+++ server/storage-jdbc/src/main/java/org/bithon/server/storage/jdbc/tracing/IndexedTagQueryBuilder.java
@@ -1,0 +1,82 @@
+/*
+ *    Copyright 2020 bithon.org
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+
+package org.bithon.server.storage.jdbc.tracing;
+
+import org.bithon.component.commons.utils.StringUtils;
+import org.bithon.server.storage.datasource.DataSourceSchema;
+import org.bithon.server.storage.jdbc.jooq.Tables;
+import org.bithon.server.storage.jdbc.utils.SQLFilterBuilder;
+import org.bithon.server.storage.metrics.DimensionFilter;
+import org.bithon.server.storage.metrics.IFilter;
+import org.bithon.server.storage.tracing.index.TagIndexConfig;
+import org.jooq.Record1;
+import org.jooq.SelectConditionStep;
+
+import java.util.Map;
+
+/**
+ * @author frank.chen021@outlook.com
+ * @date 2023/5/7 15:42
+ */
+class IndexedTagQueryBuilder extends NestQueryBuilder {
+
+    private final DataSourceSchema traceTagIndexSchema;
+
+    IndexedTagQueryBuilder(DataSourceSchema traceTagIndexSchema) {
+        this.traceTagIndexSchema = traceTagIndexSchema;
+    }
+
+    @Override
+    public SelectConditionStep<Record1<String>> build(Map<Integer, IFilter> filters) {
+        if (filters.isEmpty()) {
+            return null;
+        }
+
+        SelectConditionStep<Record1<String>> query = null;
+
+        for (Map.Entry<Integer, IFilter> entry : filters.entrySet()) {
+            Integer index = entry.getKey();
+            IFilter filter = entry.getValue();
+
+            if (index > Tables.BITHON_TRACE_SPAN_TAG_INDEX.fieldsRow().size() - 2) {
+                throw new RuntimeException(StringUtils.format("Tag [%s] is configured to use wrong index [%d]. Should be in the range [1, %d]",
+                                                              filter.getName(),
+                                                              index,
+                                                              Tables.BITHON_TRACE_SPAN_TAG_INDEX.fieldsRow().size() - 2));
+            }
+
+            if (query == null) {
+                query = dslContext.select(Tables.BITHON_TRACE_SPAN_TAG_INDEX.TRACEID)
+                                  .from(Tables.BITHON_TRACE_SPAN_TAG_INDEX)
+                                  .where(Tables.BITHON_TRACE_SPAN_TAG_INDEX.TIMESTAMP.ge(this.start))
+                                  .and(Tables.BITHON_TRACE_SPAN_TAG_INDEX.TIMESTAMP.lt(this.end));
+            }
+            query = query.and(filter.getMatcher().accept(new SQLFilterBuilder(this.traceTagIndexSchema,
+                                                                              new DimensionFilter("f" + index, filter.getMatcher()))));
+        }
+
+        if (query != null) {
+            if (this.in != null) {
+                return query.and(Tables.BITHON_TRACE_SPAN.TRACEID.in(this.in));
+            } else {
+                return query;
+            }
+        } else {
+            return this.in;
+        }
+    }
+}
