+++ server/web-service/src/main/java/org/bithon/server/web/service/tracing/service/TraceService.java
@@ -16,20 +16,14 @@
 
 package org.bithon.server.web.service.tracing.service;
 
-import org.bithon.component.commons.expression.BinaryExpression;
-import org.bithon.component.commons.expression.IExpression;
-import org.bithon.component.commons.expression.IExpressionVisitor;
-import org.bithon.component.commons.expression.IdentifierExpression;
-import org.bithon.component.commons.expression.LiteralExpression;
-import org.bithon.component.commons.expression.LogicalExpression;
-import org.bithon.server.commons.matcher.StringEqualMatcher;
+import org.bithon.component.commons.utils.StringUtils;
 import org.bithon.server.commons.time.TimeSpan;
-import org.bithon.server.storage.common.expression.FilterExpressionASTFactory;
-import org.bithon.server.storage.metrics.DimensionFilter;
 import org.bithon.server.storage.metrics.IFilter;
 import org.bithon.server.storage.tracing.ITraceReader;
 import org.bithon.server.storage.tracing.ITraceStorage;
 import org.bithon.server.storage.tracing.TraceSpan;
+import org.bithon.server.storage.tracing.TraceStorageConfig;
+import org.bithon.server.storage.tracing.index.TagIndexConfig;
 import org.bithon.server.storage.tracing.mapping.TraceIdMapping;
 import org.bithon.server.web.service.WebServiceModuleEnabler;
 import org.bithon.server.web.service.datasource.api.TimeSeriesQueryResult;
@@ -37,7 +31,6 @@ import org.bithon.server.web.service.tracing.api.TraceSpanBo;
 import org.springframework.beans.BeanUtils;
 import org.springframework.context.annotation.Conditional;
 import org.springframework.stereotype.Service;
-import org.springframework.util.StringUtils;
 
 import java.sql.Timestamp;
 import java.util.ArrayList;
@@ -45,6 +38,7 @@ import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
+import java.util.TreeMap;
 import java.util.stream.Collectors;
 
 /**
@@ -56,9 +50,12 @@ import java.util.stream.Collectors;
 public class TraceService {
 
     private final ITraceReader traceReader;
+    private final TraceStorageConfig traceStorageConfig;
 
-    public TraceService(ITraceStorage traceStorage) {
+    public TraceService(ITraceStorage traceStorage,
+                        TraceStorageConfig traceStorageConfig) {
         this.traceReader = traceStorage.createReader();
+        this.traceStorageConfig = traceStorageConfig;
     }
 
     /**
@@ -158,7 +155,7 @@ public class TraceService {
                 if (parentSpan == null) {
                     // For example, two applications: A --> B
                     // if span logs of A are not stored in Bithon,
-                    // the root span of B has parentSpanId but of course has no parent span
+                    // the root span of B has parentSpanId, but apparently the parent span can't be found
                     rootSpans.add(bo);
                 } else {
                     parentSpan.children.add(bo);
@@ -168,18 +165,36 @@ public class TraceService {
         return rootSpans;
     }
 
-    public int getTraceListSize(List<IFilter> filters, Timestamp start, Timestamp end) {
-        return traceReader.getTraceListSize(filters, start, end);
+    public int getTraceListSize(List<IFilter> filters,
+                                String filterExpression,
+                                Timestamp start,
+                                Timestamp end) {
+        // Convert filter expression into filter objects first
+        filters.addAll(FilterExpressionToFilters.toFilter(filterExpression));
+
+        FilterSplitter splitter = new FilterSplitter(this.traceStorageConfig);
+        splitter.split(filters);
+
+        return traceReader.getTraceListSize(splitter.filters, splitter.indexedTagFilter, splitter.nonIndexedTagFilter, start, end);
     }
 
     public List<TraceSpan> getTraceList(List<IFilter> filters,
+                                        String filterExpression,
                                         Timestamp start,
                                         Timestamp end,
                                         String orderBy,
                                         String order,
                                         int pageNumber,
                                         int pageSize) {
-        return traceReader.getTraceList(filters,
+        // Convert filter expression into filter objects first
+        filters.addAll(FilterExpressionToFilters.toFilter(filterExpression));
+
+        FilterSplitter splitter = new FilterSplitter(this.traceStorageConfig);
+        splitter.split(filters);
+
+        return traceReader.getTraceList(splitter.filters,
+                                        splitter.indexedTagFilter,
+                                        splitter.nonIndexedTagFilter,
                                         start,
                                         end,
                                         orderBy,
@@ -192,72 +207,60 @@ public class TraceService {
                                                       TimeSpan start,
                                                       TimeSpan end,
                                                       int bucketCount) {
-        if (StringUtils.hasText(filterExpression)) {
-            IExpression expressionAST = FilterExpressionASTFactory.create(filterExpression);
-            expressionAST.accept(new IExpressionVisitor<Void>() {
-                @Override
-                public Void visit(LogicalExpression expression) {
-                    if (expression instanceof LogicalExpression.OR) {
-                        throw new UnsupportedOperationException("OR operator is not supported now.");
-                    }
-                    for (IExpression operand : expression.getOperands()) {
-                        operand.accept(this);
-                    }
-                    return null;
-                }
+        // Convert filter expression into filter objects first
+        filters.addAll(FilterExpressionToFilters.toFilter(filterExpression));
 
-                @Override
-                public Void visit(BinaryExpression.EQ expression) {
-                    IExpression left = expression.getLeft();
-                    if (!(left instanceof IdentifierExpression)) {
-                        throw new UnsupportedOperationException("Expression at left side must be a field");
-                    }
-
-                    IExpression right = expression.getRight();
-                    if (!(right instanceof LiteralExpression)) {
-                        throw new UnsupportedOperationException("Expression at right side must be a constant");
-                    }
-
-                    filters.add(new DimensionFilter(((IdentifierExpression) left).getIdentifier(),
-                                                    new StringEqualMatcher((String) ((LiteralExpression) right).getValue())));
-                    return null;
-                }
-
-                @Override
-                public Void visit(BinaryExpression.GT expression) {
-                    return IExpressionVisitor.super.visit(expression);
-                }
-
-                @Override
-                public Void visit(BinaryExpression.GTE expression) {
-                    return IExpressionVisitor.super.visit(expression);
-                }
-
-                @Override
-                public Void visit(BinaryExpression.LT expression) {
-                    return IExpressionVisitor.super.visit(expression);
-                }
-
-                @Override
-                public Void visit(BinaryExpression.LTE expression) {
-                    return IExpressionVisitor.super.visit(expression);
-                }
-
-                @Override
-                public Void visit(BinaryExpression.NE expression) {
-                    return IExpressionVisitor.super.visit(expression);
-                }
-            });
-        }
+        FilterSplitter splitter = new FilterSplitter(this.traceStorageConfig);
+        splitter.split(filters);
 
         int bucketInSecond = getTimeBucket(start.getMilliseconds(), end.getMilliseconds(), bucketCount).length;
-        List<Map<String, Object>> dataPoints = traceReader.getTraceDistribution(filters,
+        List<Map<String, Object>> dataPoints = traceReader.getTraceDistribution(splitter.filters,
+                                                                                splitter.indexedTagFilter,
+                                                                                splitter.nonIndexedTagFilter,
                                                                                 start.toTimestamp(),
                                                                                 end.toTimestamp(),
                                                                                 bucketInSecond);
         List<String> metrics = Arrays.asList("count", "minResponse", "avgResponse", "maxResponse");
         return TimeSeriesQueryResult.build(start, end, bucketInSecond, dataPoints, "_timestamp", Collections.emptyList(), metrics);
     }
+
+    static class FilterSplitter {
+        private final List<IFilter> filters;
+        private final Map<Integer, IFilter> indexedTagFilter;
+        private final List<IFilter> nonIndexedTagFilter;
+        private final TraceStorageConfig traceStorageConfig;
+
+        public FilterSplitter(TraceStorageConfig traceStorageConfig) {
+            this.traceStorageConfig = traceStorageConfig;
+            this.filters = new ArrayList<>();
+            this.indexedTagFilter = new TreeMap<>();
+            this.nonIndexedTagFilter = new ArrayList<>();
+        }
+
+        void split(List<IFilter> filters) {
+            TagIndexConfig indexedTagConfig = this.traceStorageConfig.getIndexes();
+
+            for (IFilter filter : filters) {
+                if (!filter.getName().startsWith("tags.")) {
+                    this.filters.add(filter);
+                    continue;
+                }
+
+                String tagName = filter.getName().substring("tags".length());
+                if (!StringUtils.hasText(tagName)) {
+                    throw new RuntimeException(StringUtils.format("Wrong tag name [%s]", filter.getName()));
+                }
+
+                int tagIndex = indexedTagConfig == null ? 0 : indexedTagConfig.getColumnPos(tagName);
+                if (tagIndex > 0) {
+                    indexedTagFilter.put(tagIndex, filter);
+                } else {
+                    nonIndexedTagFilter.add(filter);
+                }
+            }
+        }
+    }
+
 
     static class Bucket {
         /**
